Refactor clear selection button out of GeographyInfo so it can be reused
